+++ Scripts/Save/StockSave.cs
@@ -1,7 +1,7 @@
 public class Stock {
     public string Code = "BRUH";
-    public int Amount = 0;
-    public int Price = 69;
+    public long Amount = 0;
+    public long Price = 69;
 
     public Stock(string epicCode, int epicAmount, int epicPrice) {
         Code = epicCode;
