+++ Scripts/UI/Advisors/LawsuitOK.cs
@@ -7,7 +7,7 @@ public class LawsuitOK : TextureButton {
     }
 
     public void Click() {
-        Global.Money += (int)Global.Money/20;
+        Global.Money += (long)Global.Money/20;
         if (GetNodeOrNull("/root/PauseThingy") != null)
             GetNode("/root/PauseThingy").QueueFree();
         GetParent().QueueFree();
