+++ Scripts/UI/NewProduct/Food/FoodPriceCalculator.cs
@@ -39,7 +39,7 @@ public class FoodPriceCalculator : Label {
             Global.FoodNewProduct.Rating = 10;
         
         if (Global.Investors/50 > 1)
-            Global.FoodNewProduct.MoneySpent *= ((int)Global.Investors/50);
+            Global.FoodNewProduct.MoneySpent *= ((long)Global.Investors/50);
 
         Text = "This product will cost $" + String.Format("{0:n0}", Global.FoodNewProduct.MoneySpent) +
         " to develop";
